Enable ktfmt to format Kotlin code
Before this commit, I spent a while manually formatting code to reduce
diffs in the codebase. By using ktfmt, we can focus more on the code and
less on trivia like formatting. The formatted output isn't always
perfect, but it's better than not having a code formatter.

We use the Dropbox style instead of the default style because we want to
use 4-space indent instead of 2-space indent, for consistency with
the official Kotlin style guide and google-java-format (which we use in lsif-java)