Upgrade to Kotlin 1.9.x
Fixes #74. My limited understanding of Kotlin compiler compatibility is
that this change drops support for 1.8.x. Until we move to the more
stable Kotlin compiler API (which I have barely researched), I think
it's reasonable to target the latest Kotlin version in scip-kotlin.
Users on older versions of Kotlin will need to stay on older versions of
scip-kotlin.